move useRoomInfo hook
