+++ packages/react/src/components/RoomName.tsx
@@ -1,26 +1,5 @@
-import { roomInfoObserver } from '@livekit/components-core';
-import type { Room } from 'livekit-client';
 import * as React from 'react';
-
-import { useEnsureRoom } from '../context';
-import { useObservableState } from '../hooks/internal/useObservableState';
-
-/** @public */
-export interface UseRoomInfoOptions {
-  room?: Room;
-}
-
-/** @public */
-export function useRoomInfo(options: UseRoomInfoOptions = {}) {
-  const room = useEnsureRoom(options.room);
-  const infoObserver = React.useMemo(() => roomInfoObserver(room), [room]);
-  const { name, metadata } = useObservableState(infoObserver, {
-    name: room.name,
-    metadata: room.metadata,
-  });
-
-  return { name, metadata };
-}
+import { useRoomInfo } from '../hooks';
 
 /** @public */
 export interface RoomNameProps extends React.HTMLAttributes<HTMLSpanElement> {
