+++ packages/react/src/hooks/index.ts
@@ -25,3 +25,4 @@ export { useVisualStableUpdate, UseVisualStableUpdateOptions } from './useVisual
 export { usePinnedTracks } from './usePinnedTracks';
 export { useSwipe, UseSwipeOptions } from './useSwipe';
 export { useFacingMode } from './useFacingMode';
+export { useRoomInfo } from './useRoomInfo';
